+++ f5/bigip/test/unit/test_resource.py
@@ -25,17 +25,13 @@ from f5.bigip.resource import AsmResource
 from f5.bigip.resource import AttemptedMutationOfReadOnly
 from f5.bigip.resource import BooleansToReduceHaveSameValue
 from f5.bigip.resource import Collection
-from f5.bigip.resource import GenerationMismatch
-from f5.bigip.resource import InvalidForceType
 from f5.bigip.resource import MissingRequiredCreationParameter
-from f5.bigip.resource import MissingRequiredReadParameter
 from f5.bigip.resource import OrganizingCollection
 from f5.bigip.resource import PathElement
 from f5.bigip.resource import Resource
 from f5.bigip.resource import ResourceBase
 from f5.bigip.resource import Stats
 from f5.bigip.resource import UnnamedResource
-from f5.bigip.resource import UnregisteredKind
 from f5.bigip.resource import UnsupportedOperation
 from f5.bigip.resource import URICreationCollision
 from f5.bigip.tm.asm.signature_sets import Signature_Set
@@ -46,10 +42,14 @@ from f5.bigip.tm.ltm.virtual import Profiles_s
 from f5.bigip.tm.ltm.virtual import Virtual
 from f5.sdk_exception import DeviceProvidesIncompatibleKey
 from f5.sdk_exception import ExclusiveAttributesPresent
+from f5.sdk_exception import GenerationMismatch
+from f5.sdk_exception import InvalidForceType
 from f5.sdk_exception import InvalidResource
 from f5.sdk_exception import KindTypeMismatch
 from f5.sdk_exception import MissingRequiredCommandParameter
+from f5.sdk_exception import MissingRequiredReadParameter
 from f5.sdk_exception import RequestParamKwargCollision
+from f5.sdk_exception import UnregisteredKind
 from f5.sdk_exception import UnsupportedMethod
 from icontrol.exceptions import iControlUnexpectedHTTPError
