+++ f5/bigip/tm/gtm/test/functional/test_listener.py
@@ -18,9 +18,9 @@ import pytest
 
 from distutils.version import LooseVersion
 from f5.bigip.resource import MissingRequiredCreationParameter
-from f5.bigip.resource import MissingRequiredReadParameter
 from f5.bigip.resource import UnsupportedOperation
 from f5.bigip.tm.gtm.listener import Listener
+from f5.sdk_exception import MissingRequiredReadParameter
 from pytest import symbols
 from requests.exceptions import HTTPError
 from six import iteritems
