+++ f5/bigip/tm/gtm/test/functional/test_topology.py
@@ -17,10 +17,10 @@
 import pytest
 
 from distutils.version import LooseVersion
-from f5.bigip.resource import InvalidName
 from f5.bigip.resource import MissingRequiredCreationParameter
 from f5.bigip.resource import UnsupportedOperation
 from f5.bigip.tm.gtm.topology import Topology
+from f5.sdk_exception import InvalidName
 from f5.sdk_exception import UnsupportedTmosVersion
 from pytest import symbols
 from requests.exceptions import HTTPError
