+++ f5/bigip/tm/gtm/test/unit/test_topology.py
@@ -16,11 +16,11 @@
 import pytest
 
 from f5.bigip import ManagementRoot
-from f5.bigip.resource import InvalidName
 from f5.bigip.resource import MissingRequiredCreationParameter
 from f5.bigip.resource import UnsupportedOperation
 from f5.bigip.tm import Gtm
 from f5.bigip.tm.gtm.topology import Topology
+from f5.sdk_exception import InvalidName
 
 
 @pytest.fixture
