+++ f5/bigip/tm/gtm/topology.py
@@ -28,10 +28,10 @@ REST Kind
 """
 from distutils.version import LooseVersion
 from f5.bigip.resource import Collection
-from f5.bigip.resource import InvalidName
 from f5.bigip.resource import Resource
 from f5.bigip.resource import UnsupportedOperation
 from f5.bigip.resource import URICreationCollision
+from f5.sdk_exception import InvalidName
 from f5.sdk_exception import UnsupportedTmosVersion
